add abstract class
