+++ src/Reader/Filter/LessThan.php
@@ -3,26 +3,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Data\Reader\Filter;
 
-final class LessThan implements FilterInterface
+final class LessThan extends CompareCriterion
 {
-    private $field;
-    private $value;
-
-    public function __construct(string $field, $value)
-    {
-        if (!is_scalar($value)) {
-            throw new \InvalidArgumentException('Value should be scalar');
-        }
-
-        $this->field = $field;
-        $this->value = $value;
-    }
-
-    public function toArray(): array
-    {
-        return [self::getOperator(), $this->field, $this->value];
-    }
-
     public static function getOperator(): string
     {
         return 'lt';
